+++ src/app/layout/navigation/navigation.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnInit, ViewChild, ElementRef } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { MatDialog } from '@angular/material';
 import { LocaleService, TranslationType } from '@core/locale.service';
@@ -25,6 +25,9 @@ export class NavigationComponent extends Unsubscribable implements OnInit {
     @Input()
     public user: User = null;
 
+    @ViewChild('donate')
+    public donateButton: ElementRef;
+
     public tools = [
         {
             name: 'Church Circles',
@@ -49,6 +52,8 @@ export class NavigationComponent extends Unsubscribable implements OnInit {
             .subscribe(result => {
                 this.localeService.set(result);
             });
+
+        this.donateButton.nativeElement.innerHTML = getPayPalButton();
     }
 
     public goto(event: Event, url: string): void {
@@ -76,4 +81,34 @@ export class NavigationComponent extends Unsubscribable implements OnInit {
             }
         });
     }
+}
+
+
+function getPayPalButton(): string {
+    return `
+    <form action="https://www.paypal.com/cgi-bin/webscr"
+          method="post"
+          target="_top">
+        <input type="hidden"
+               name="cmd"
+               value="_donations" />
+        <input type="hidden"
+               name="business"
+               value="WR657KF67GLC2" />
+        <input type="hidden"
+               name="currency_code"
+               value="USD" />
+        <input type="image"
+               src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif"
+               border="0"
+               name="submit"
+               title="PayPal - The safer, easier way to pay online!"
+               alt="Donate with PayPal button" />
+        <img alt=""
+             border="0"
+             src="https://www.paypal.com/en_US/i/scr/pixel.gif"
+             width="1"
+             height="1" />
+    </form>
+    `;
 }