+++ src/app/tools/gen-mapper/gen-mapper-graph/gen-mapper-graph.component.ts
@@ -20,6 +20,7 @@ import { GMTemplate, GNode } from '../gen-mapper.interface';
 import { NodeClipboardService } from '../node-clipboard.service';
 import { cloneDeep } from 'lodash';
 import { LocaleService } from '@core/locale.service';
+import { MatSnackBar } from '@angular/material';
 
 @Component({
     selector: 'app-gen-mapper-graph',
@@ -47,6 +48,7 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
         private locale: LocaleService,
         private dialog: MatDialog,
         private nodeClipboard: NodeClipboardService,
+        private snackBar: MatSnackBar,
         private elementRef: ElementRef
     ) { }
 
@@ -81,9 +83,13 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
             this._updating = false;
         };
 
-        this.graph.onCopyNode = (nodes: GNode[]) => { };
+        this.graph.onCopyNode = (nodes: GNode[]) => {
+            // Buttons on graph is not implemented
+        };
 
-        this.graph.onPasteNode = (d: any) => { };
+        this.graph.onPasteNode = (d: any) => {
+            // Buttons on graph is not implemented
+        };
 
         this.graph.removeNodeClick = (node: any) => {
             const name = node.data.name || node.data.leaderName || 'No Name';
@@ -104,11 +110,18 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
         };
 
         this.graph.nodeClick = (node: any) => {
+
+            // Setup descendants for copy/paste
+            const descendants = cloneDeep(node.descendants().map(d => d.data));
+            const same = descendants.find(n => n.id === node.data.id);
+            same.parentId = '';
+
             this.dialog
                 .open(EditNodeDialogComponent, {
                     minWidth: '400px',
                     data: {
                         nodeData: node.data,
+                        descendants: descendants,
                         template: this.template,
                         language: this.graph.language,
                         nodes: this.graph.data
@@ -120,6 +133,12 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
                         return;
                     }
 
+                    if (result.isPasteNode) {
+                        const originalData = cloneDeep(this.graph.data);
+                        this.graph.pasteNode(node, this.nodeClipboard.getValue());
+                        this.showUndoPaste(originalData);
+                    }
+
                     if (result.isImportSubtree) {
                         this.graph.csvIntoNode(node, result.content);
                     }
@@ -129,5 +148,17 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
                     }
                 });
         };
+    }
+
+    private showUndoPaste(originalData: GNode[]): void {
+        this.snackBar
+            .open(this.locale.t('nodeHasBeenReplaces'), this.locale.t('en_Undo'), {
+                duration: 20000,
+            })
+            .onAction()
+            .pipe(take(1))
+            .subscribe(result => {
+                this.graph.redrawData(originalData);
+            });
     }
 }